+++ main/managers/views/nfc_view.c
@@ -5994,6 +5994,7 @@ void nfc_view_create(void) {
     lv_obj_clear_flag(touch_bar, LV_OBJ_FLAG_SCROLLABLE | LV_OBJ_FLAG_CLICKABLE);
 
     scroll_up_btn = lv_btn_create(touch_bar);
+    gui_apply_pressed_style(scroll_up_btn);
     lv_obj_set_size(scroll_up_btn, SCROLL_BTN_SIZE, SCROLL_BTN_SIZE);
     lv_obj_align(scroll_up_btn, LV_ALIGN_LEFT_MID, SCROLL_BTN_PADDING, 0);
     lv_obj_set_style_bg_color(scroll_up_btn, ctrl_color, LV_PART_MAIN);
@@ -6008,6 +6009,7 @@ void nfc_view_create(void) {
     lv_obj_add_flag(scroll_up_btn, LV_OBJ_FLAG_HIDDEN);
 
     back_btn = lv_btn_create(touch_bar);
+    gui_apply_pressed_style(back_btn);
     lv_obj_set_size(back_btn, SCROLL_BTN_SIZE + 24, SCROLL_BTN_SIZE);
     lv_obj_align(back_btn, LV_ALIGN_CENTER, 0, 0);
     lv_obj_set_style_bg_color(back_btn, ctrl_color, LV_PART_MAIN);
@@ -6022,6 +6024,7 @@ void nfc_view_create(void) {
     lv_obj_center(back_label);
 
     scroll_down_btn = lv_btn_create(touch_bar);
+    gui_apply_pressed_style(scroll_down_btn);
     lv_obj_set_size(scroll_down_btn, SCROLL_BTN_SIZE, SCROLL_BTN_SIZE);
     lv_obj_align(scroll_down_btn, LV_ALIGN_RIGHT_MID, -SCROLL_BTN_PADDING, 0);
     lv_obj_set_style_bg_color(scroll_down_btn, ctrl_color, LV_PART_MAIN);